+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.2.0] - Change internal API.
+- From `interface` to `abstract class`.
+- Add [getDyLibModule] code docs to read API easily.
+
+## [0.1.0] - Change internal API.
+- Support MacOS Release Mode.
+
 ## [0.0.3] - Support Linux.
 Linux test.